Rename sequencer to writer in db modules
