+++ crates/btcio/src/writer/bundler.rs
@@ -21,7 +21,7 @@ pub(crate) async fn bundler_task(ops: Arc<EnvelopeDataOps>) -> anyhow::Result<()
 }
 
 /// Processes and bundles a list of unbundled intents into payload entries.
-/// NOTE: The logic current is simply 1-1 mapping between intents and payloads, in future it can
+/// NOTE: The current logic is simply 1-1 mapping between intents and payloads, in future it can
 /// be sophisticated.
 async fn process_unbundled_entries(
     ops: &EnvelopeDataOps,
@@ -52,30 +52,32 @@ async fn get_unbundled_intents_after(
     ops: &EnvelopeDataOps,
 ) -> anyhow::Result<(Vec<IntentEntry>, u64)> {
     let latest_idx = ops.get_next_intent_idx_async().await?.saturating_sub(1);
-    let mut curr_intent_idx = latest_idx;
-    let mut unbundled_intents = Vec::new();
-    while curr_intent_idx >= idx {
-        if let Some(intent_entry) = ops.get_intent_by_idx_async(curr_intent_idx).await? {
-            match intent_entry.status {
-                IntentStatus::Unbundled => unbundled_intents.push(intent_entry),
+    let mut curr_idx = latest_idx;
+
+    let mut unbundled = Vec::new();
+
+    while curr_idx >= idx {
+        if let Some(intent) = ops.get_intent_by_idx_async(curr_idx).await? {
+            match intent.status {
+                IntentStatus::Unbundled => unbundled.push(intent),
                 IntentStatus::Bundled(_) => {
                     // Bundled intent found, no more to scan
                     break;
                 }
             }
         } else {
-            warn!(%curr_intent_idx, "Could not find expected intent in db");
+            warn!(%curr_idx, "Could not find expected intent in db");
             break;
         }
 
-        if curr_intent_idx == 0 {
+        if curr_idx == 0 {
             break;
         }
-        curr_intent_idx -= 1;
+        curr_idx -= 1;
     }
 
     // Reverse the items so that they are in ascending order of index
-    unbundled_intents.reverse();
+    unbundled.reverse();
 
-    Ok((unbundled_intents, latest_idx))
+    Ok((unbundled, latest_idx))
 }
