+++ crates/storage/src/ops/mod.rs
@@ -3,7 +3,7 @@ pub mod bridge_duty;
 pub mod bridge_duty_index;
 pub mod bridge_relay;
 pub mod checkpoint;
-pub mod writer;
 pub mod l1;
 pub mod l1tx_broadcast;
 pub mod l2;
+pub mod writer;
